[Home] Add HomePageArtworkModule(s) to replace HomePageModule(s).
But for now HomePageModule(s) will continue to work from the root fields
`home_page_module` and `home_page_modules`.
