+++ schema/home/index.js
@@ -1,5 +1,5 @@
-import HomePageModules from './home_page_modules';
-import HomePageModule from './home_page_module';
+import HomePageArtworkModule from './home_page_artwork_module';
+import HomePageArtworkModules from './home_page_artwork_modules';
 import HomePageArtistModule from './home_page_artist_module';
 import HomePageArtistModules from './home_page_artist_modules';
 
@@ -10,8 +10,8 @@ import {
 const HomePageType = new GraphQLObjectType({
   name: 'HomePage',
   fields: {
-    artwork_module: HomePageModule,
-    artwork_modules: HomePageModules,
+    artwork_module: HomePageArtworkModule,
+    artwork_modules: HomePageArtworkModules,
     artist_module: HomePageArtistModule,
     artist_modules: HomePageArtistModules,
   },
